projects/example/requirements.txt
facebook_sdk==3.1.0
discord.py==1.2.3
tweepy==3.8.0
discord==1.0.1
python-dotenv==0.12.0
